# RecipeWebAPI/backend/docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Recipe API
  description: A simple API to manage recipes
  contact:
    name: Mustafa Safi
    email: safim4@coventry.ac.uk
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'

servers:
  - url: https://forumdispute-royalshallow-3000.codio-box.uk
    description: Development server

tags:
  - name: Recipes
    description: All of the endpoints for managing recipes

security:
  - basicAuth: []

paths:
  /recipes:
    get:
      tags:
        - Recipes
      summary: Get all recipes
      responses:
        '200':
          description: A list of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        default:
          description: Unexpected error

    post:
      tags:
        - Recipes
      summary: Create a new recipe
      requestBody:
        description: Recipe to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '201':
          description: Recipe created
        '400':
          description: Invalid input
        default:
          description: Unexpected error

  /recipes/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric ID of the recipe

    get:
      tags:
        - Recipes
      summary: Get a recipe by ID
      responses:
        '200':
          description: A single recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found
        default:
          description: Unexpected error

    put:
      tags:
        - Recipes
      summary: Update a recipe
      requestBody:
        description: Updated recipe data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: Recipe updated
        '404':
          description: Recipe not found
        default:
          description: Unexpected error

    delete:
      tags:
        - Recipes
      summary: Delete a recipe
      responses:
        '200':
          description: Recipe deleted
        '404':
          description: Recipe not found
        default:
          description: Unexpected error

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    Recipe:
      title: Recipe
      description: A recipe object
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        ingredients:
          type: string
          minLength: 1
        instructions:
          type: string
          minLength: 1
        created_by:
          type: string
          minLength: 1
        prep_time:
          type: number
          minimum: 0
        cook_time:
          type: number
          minimum: 0
      required:
        - title
        - ingredients
        - instructions
        - created_by
      additionalProperties: false